fix: fix Accordion.Body loading issue
Added an Accordion.SetLoaded component that when rendered will trigger Accordion.Body to unhide.
Added a respectSetLoaded prop to Accordion.Item for specifying whether Accordion.SetLoaded component is to be used or not.
Passed Accordion.SetLoaded component as a prop to SandboxContainer to be rendered once content.html exists.
Minor styling changes to Sandbox.Doc.